📦 feat: composition by creating box component
