+++ src/App.jsx
@@ -5,8 +5,9 @@ import Logo from './components/Nav/NavChildrens/Logo';
 import SearchInput from './components/Nav/NavChildrens/SearchInput';
 import NumResult from './components/Nav/NavChildrens/NumResult';
 
-import MovieList from './components/Main/MainChildren/MovieList';
-import WatchedMovie from './components/Main/MainChildren/WatchedMovie';
+import Box from './components/Main/MainChildren/Box';
+import WatchedSummary from './components/Main/MainChildren/WatchedSummary/WatchedSummary';
+import WatchedMoviesList from './components/Main/MainChildren/MovieListStructure/MovieStructure/WatchedList/WatchedMoviesList';
 import MovieListStructure from './components/Main/MainChildren/MovieListStructure/MovieListStructure';
 
 import { useState } from 'react';
@@ -35,8 +36,32 @@ const tempMovieData = [
   },
 ];
 
+const tempWatchedData = [
+  {
+    imdbID: 'tt1375666',
+    Title: 'Inception',
+    Year: '2010',
+    Poster:
+      'https://m.media-amazon.com/images/M/MV5BMjAxMzY3NjcxNF5BMl5BanBnXkFtZTcwNTI5OTM0Mw@@._V1_SX300.jpg',
+    runtime: 148,
+    imdbRating: 8.8,
+    userRating: 10,
+  },
+  {
+    imdbID: 'tt0088763',
+    Title: 'Back to the Future',
+    Year: '1985',
+    Poster:
+      'https://m.media-amazon.com/images/M/MV5BZmU0M2Y1OGUtZjIxNi00ZjBkLTg1MjgtOWIyNThiZWIwYjRiXkEyXkFqcGdeQXVyMTQxNzMzNDI@._V1_SX300.jpg',
+    runtime: 116,
+    imdbRating: 8.5,
+    userRating: 9,
+  },
+];
+
 export default function App() {
   const [movies, setMovies] = useState(tempMovieData);
+  const [watched, setWatched] = useState(tempWatchedData);
   return (
     <>
       <Nav>
@@ -46,11 +71,15 @@ export default function App() {
       </Nav>
 
       <Main>
-        <MovieList>
+        <Box>
           <MovieListStructure movies={movies} />
-        </MovieList>
+        </Box>
 
-        <WatchedMovie />
+        <Box>
+          <WatchedSummary watched={watched} />
+
+          <WatchedMoviesList watched={watched} />
+        </Box>
       </Main>
     </>
   );
