+++ src/components/Main/MainChildren/Box.jsx
@@ -1,0 +1,16 @@
+import React, { useState } from 'react';
+
+const Box = ({ children }) => {
+  const [isOpen, setIsOpen] = useState(true);
+
+  return (
+    <div className='box'>
+      <button className='btn-toggle' onClick={() => setIsOpen((open) => !open)}>
+        {isOpen ? '–' : '+'}
+      </button>
+      {isOpen && children}
+    </div>
+  );
+};
+
+export default Box;
