--- src/components/Main/MainChildren/MovieList.jsx
@@ -1,19 +1,0 @@
-import React, { useState } from 'react';
-
-const MovieList = ({ children }) => {
-  const [isOpen1, setIsOpen1] = useState(true);
-
-  return (
-    <div className='box'>
-      <button
-        className='btn-toggle'
-        onClick={() => setIsOpen1((open) => !open)}
-      >
-        {isOpen1 ? '–' : '+'}
-      </button>
-      {isOpen1 && children}
-    </div>
-  );
-};
-
-export default MovieList;
